\name{get.tumour.normal.ratio}
\alias{get.tumour.normal.ratio}
\title{Calculate tumour ratios}
\description{Calculates tumour ratios from samples stored in microarray chips}
\usage{get.tumour.normal.ratio(ref, nchips, chips.annot, ns.counts, output)}
\arguments{
    \item{ref}{The reference sample}
    \item{nchips}{The number of chips}
    \item{chips.annot}{Data frame comprising of the samples' annotations}
    \item{ns.counts}{The data associated with the samples}
    \item{output}{The data frame in which the results will be returned}
}
\details{None}
\value{None}
\author{Emilie Lalonde}
    

