\name{make.positive.control.plot}
\alias{make.positive.control.plot}
\title{Make heatmaps of correlations}
\description{Make heatmaps of correlations between target concentration and actual counts}
\usage{
make.positive.control.plot(correlations)
}
\arguments{
  \item{correlations}{A vector of correlations}
}
\details{Make heatmaps of correlations between target concentration and actual counts. First heatmap shows full range of correlation, and second is zoomed in to observed values only}
\value{None}
\author{Cindy Yao and Emilie Lalonde}
